+++ .tmp/push-test-app/.prettierrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-  tabWidth: 2,
-  printWidth: 140,
-  arrowParens: 'avoid',
-  bracketSameLine: false,
-  bracketSpacing: true,
-  singleQuote: false,
-  trailingComma: 'all',
-};