+++ .tmp/push-test-app/ReactotronConfig.js
@@ -1,3 +1,0 @@
-import Reactotron from "reactotron-react-native";
-
-export const reactotron = Reactotron.configure({ name: "Trofi" }).useReactNative().connect();